Updates to allow for default 'byId' methods

+++ src/collection.utils.ts
@@ -5,36 +5,40 @@ import { IMakeKey } from '.';
 const logger = Logger.make('CollectionUtils');
 
 const _Array = {
-  getKeys: <T>({ collection, makeKey }: { collection: Array<T>; makeKey: IMakeKey<T> }) => collection.map((item) => makeKey(item)),
-  getItem: <T>({ collection, makeKey, key }: { collection: Array<T>; makeKey: IMakeKey<T>; key: string }) =>
-    collection.find((item) => {
-      return makeKey(item) === key;
-    }),
-  insertItem: <T>({ collection, key, value }: { collection: Array<T>; key: string; value: T }) => {
-    collection.push(value);
+  getKeys: <T>({ collection, makeKey }: { collection: Array<T>; makeKey: IMakeKey<T> }) => {
+    return collection.length > 0 ? collection.map((item) => makeKey(item)) : [];
   },
+  getItem: <T>({ collection, makeKey, key }: { collection: Array<T>; makeKey: IMakeKey<T>; key: string }) => {
+    return collection.length > 0 ? collection.find((item) => makeKey(item) === key) : undefined;
+  },
+  insertItem: <T>({ collection, key, value }: { collection: Array<T>; key: string; value: T }) => collection.push(value),
   updateItem: <T>({ collection, makeKey, value }: { collection: Array<T>; makeKey: IMakeKey<T>; value: T }) => {
+    if (collection.length === 0) return;
     const key = makeKey(value);
     const existingItemIndex = collection.findIndex((item) => makeKey(item) === key);
     if (existingItemIndex) {
       collection.splice(existingItemIndex, 1, value);
     }
   },
-  deleteItem: <T>({ collection, makeKey, key }: { collection: Array<T>; makeKey: IMakeKey<T>; key: string }) =>
+  deleteItem: <T>({ collection, makeKey, key }: { collection: Array<T>; makeKey: IMakeKey<T>; key: string }) => {
+    if (collection.length === 0) return;
     collection.splice(
       collection.findIndex((item) => makeKey(item) === key),
       1,
-    ),
+    );
+  },
   clear: <T>({ collection }: { collection: Array<T> }) => collection.splice(0, collection.length),
 };
 
 const _Set = {
-  getKeys: <T>({ collection, makeKey }: { collection: Set<T>; makeKey: IMakeKey<T> }) => Array.from(collection.values()).map((domainItem) => makeKey(domainItem)),
-  getItem: <T>({ collection, makeKey, key }: { collection: Set<T>; makeKey: IMakeKey<T>; key: string }) => Array.from(collection.values()).find((domainItem) => makeKey(domainItem) == key),
+  getKeys: <T>({ collection, makeKey }: { collection: Set<T>; makeKey: IMakeKey<T> }) => (collection.size > 0 ? Array.from(collection.values()).map((domainItem) => makeKey(domainItem)) : []),
+  tryGetItem: <T>({ collection, makeKey, key }: { collection: Set<T>; makeKey: IMakeKey<T>; key: string }) =>
+    collection.size > 0 ? Array.from(collection.values()).find((domainItem) => makeKey(domainItem) == key) : undefined,
   insertItem: <T>({ collection, key, value }: { collection: Set<T>; key: string; value: T }) => {
     collection.add(value);
   },
-  updateItem: <T>({ collection, makeKey, value }: { collection: Set<T>; makeKey: IMakeKey<T>; value: T }) => {
+  tryUpdateItem: <T>({ collection, makeKey, value }: { collection: Set<T>; makeKey: IMakeKey<T>; value: T }) => {
+    if (collection.size === 0) return;
     const key = makeKey(value);
     const existingItem = Array.from(collection.values()).find((domainItem) => makeKey(domainItem) === key);
     if (existingItem) {
@@ -42,7 +46,8 @@ const _Set = {
     }
     collection.add(value);
   },
-  deleteItem: <T>({ collection, makeKey, key }: { collection: Set<T>; makeKey: IMakeKey<T>; key: string }) => {
+  tryDeleteItem: <T>({ collection, makeKey, key }: { collection: Set<T>; makeKey: IMakeKey<T>; key: string }) => {
+    if (collection.size === 0) return;
     const item = Array.from(collection.values()).find((domainItem) => makeKey(domainItem) === key);
     if (item) collection.delete(item);
   },
@@ -50,14 +55,14 @@ const _Set = {
 
 const _Record = {
   getKeys: <T>({ collection }: { collection: Record<string, TextDecodeOptions> }) => Object.keys(collection),
-  getItem: <T>({ collection, key }: { collection: Record<string, T>; key: string }) => collection[key],
+  tryGetItem: <T>({ collection, key }: { collection: Record<string, T>; key: string }) => collection[key],
   insertItem: <T>({ collection, key, value }: { collection: Record<string, T>; key: string; value: T }) => {
     collection[key] = value;
   },
-  updateItem: <T>({ collection, key, value }: { collection: Record<string, T>; key: string; value: T }) => {
+  tryUpdateItem: <T>({ collection, key, value }: { collection: Record<string, T>; key: string; value: T }) => {
     collection[key] = value;
   },
-  deleteItem: <T>({ collection, key }: { collection: Record<string, T>; key: string }) => {
+  tryDeleteItem: <T>({ collection, key }: { collection: Record<string, T>; key: string }) => {
     delete collection[key];
   },
 };
